refactor(community): CommunityBoardRepository 네이밍 변경
